+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Popmotion adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.1.0] 2015-11-28
+
+### Added
+- Actor value `watch` property can now be a function returning any numerical value.
+- Added `smooth` for all actions.
+- Added `calc.toDecimal` calculator function. 
+
 ## [4.0.0] 2015-11-20
 
 ### Added